Migrate project pages (Task 2.4) to App Router
