+++ src/app/projects/list/page.js
@@ -1,0 +1,5 @@
+import { redirect } from "next/navigation";
+
+export default function ProjectsListIndexPage() {
+  redirect("/projects/list/1");
+}
